Moved components (#20)
